Shrink EdgeSourceInfo

+++ include/geos/operation/overlayng/EdgeSourceInfo.h
@@ -41,7 +41,7 @@ private:
 
     // Members
     uint8_t index;
-    int dim;
+    int8_t dim;
     bool edgeIsHole;
     int depthDelta;
